fix some issues + small refactor
