+++ web_src/src/pages/workflowv2/index.tsx
@@ -124,6 +124,7 @@ import {
   clearComponentSidebarSearchParams,
   getExitEditModeDisabledTooltip,
   getRunActionState,
+  getWorkflowViewFlagsFromSearchParams,
   getWorkflowViewPresentation,
   readStoredBoolean,
 } from "./viewState";
@@ -197,6 +198,7 @@ export function WorkflowPageV2() {
     selectedRunId,
     setSelectedRunId,
   } = useWorkflowViewSearchParams(searchParams, setSearchParams);
+  const urlViewFlags = useMemo(() => getWorkflowViewFlagsFromSearchParams(searchParams), [searchParams]);
   const currentUserId = me?.id;
   const { canAct } = usePermissions();
   const [activeCanvasVersion, setActiveCanvasVersion] = useState<CanvasesCanvasVersion | null>(null);
@@ -4914,10 +4916,10 @@ export function WorkflowPageV2() {
     onDashboardOpenYaml,
     dashboardYamlReadOnly,
   } = useWorkflowViewModeActions({
-    isDashboardMode,
-    isMemoryMode,
-    isFilesMode,
-    isRunsMode,
+    isDashboardMode: urlViewFlags.isDashboardMode,
+    isMemoryMode: urlViewFlags.isMemoryMode,
+    isFilesMode: urlViewFlags.isFilesMode,
+    isRunsMode: urlViewFlags.isRunsMode,
     hasEditableVersion,
     isTemplate,
     canUpdateCanvas,
@@ -5476,10 +5478,7 @@ export function WorkflowPageV2() {
     hasDraftDiffVersusLive: !!latestDraftVersion && hasDraftDiffVersusLive,
   });
   const { headerMode, canvasStateMode } = getWorkflowViewPresentation({
-    isDashboardMode,
-    isRunsMode,
-    isMemoryMode,
-    isFilesMode,
+    ...urlViewFlags,
     hasEditableVersion,
     isViewingPendingApprovalVersion,
     isViewingCurrentLiveVersion,
@@ -5500,7 +5499,7 @@ export function WorkflowPageV2() {
   runDisabledRef.current = runDisabled;
   runDisabledTooltipRef.current = runDisabledTooltip;
   const filesHeaderActionsSlotId = canvasId ? `canvas-files-header-actions-${canvasId}` : "canvas-files-header-actions";
-  const useFilesHeaderActions = isFilesMode && isEditing;
+  const useFilesHeaderActions = urlViewFlags.isFilesMode && isEditing;
   const resolvedPublishVersionHandler = useFilesHeaderActions
     ? filesHeaderActions?.onPublish
     : isChangeManagementDisabled
@@ -5531,7 +5530,7 @@ export function WorkflowPageV2() {
     <>
       <div className="relative h-full w-full">
         <WorkflowDashboardOverlay
-          isDashboardMode={isDashboardMode}
+          isDashboardMode={urlViewFlags.isDashboardMode}
           canActOnCanvas={canActOnCanvas}
           editLocked={isReadOnly}
           showDashboardEditControls={isEditing}
@@ -5552,7 +5551,7 @@ export function WorkflowPageV2() {
           onTriggerNode={handleDashboardTriggerNode}
         />
         <WorkflowMemoryOverlayLayer
-          isMemoryMode={isMemoryMode}
+          isMemoryMode={urlViewFlags.isMemoryMode}
           canDelete={canEditCanvasMemory({
             ...canvasAccess,
             hasEditableVersion,
@@ -5562,15 +5561,17 @@ export function WorkflowPageV2() {
           error={canvasMemoryError}
           deleteCanvasMemoryEntry={deleteCanvasMemoryEntry}
         />
-        <WorkflowFilesOverlayLayer
-          isFilesMode={isFilesMode}
-          isEditing={isEditing}
-          canvasId={canvasId || undefined}
-          canWrite={canActOnCanvas}
-          files={workflowFiles}
-          headerActionsSlotId={filesHeaderActionsSlotId}
-          onHeaderActionsChange={handleFilesHeaderActionsChange}
-        />
+        {urlViewFlags.isFilesMode ? (
+          <WorkflowFilesOverlayLayer
+            isFilesMode
+            isEditing={isEditing}
+            canvasId={canvasId || undefined}
+            canWrite={canActOnCanvas}
+            files={workflowFiles}
+            headerActionsSlotId={filesHeaderActionsSlotId}
+            onHeaderActionsChange={handleFilesHeaderActionsChange}
+          />
+        ) : null}
         <CanvasPage
           key={canvasRenderKey}
           // Persist right sidebar in query params
@@ -5591,8 +5592,12 @@ export function WorkflowPageV2() {
           hasAutoOpenedVersionControl={hasAutoOpenedVersionControl}
           onVersionControlAutoOpened={handleVersionControlAutoOpened}
           onCloseVersionControl={handleCloseVersionControl}
-          showBottomStatusControls={!isTemplate && !isRunsMode && !isMemoryMode && !isFilesMode}
-          hideAddControls={isTemplate || isRunsMode || isMemoryMode || isFilesMode}
+          showBottomStatusControls={
+            !isTemplate && !urlViewFlags.isRunsMode && !urlViewFlags.isMemoryMode && !urlViewFlags.isFilesMode
+          }
+          hideAddControls={
+            isTemplate || urlViewFlags.isRunsMode || urlViewFlags.isMemoryMode || urlViewFlags.isFilesMode
+          }
           hideCanvasToolSidebar={isTemplate}
           onSelectMemory={isTemplate ? undefined : handleSelectMemoryMode}
           nodes={nodes}
@@ -5636,7 +5641,7 @@ export function WorkflowPageV2() {
           canUpdateIntegrations={canUpdateIntegrations}
           missingIntegrations={missingIntegrations}
           onConnectIntegration={!isReadOnly ? handleConnectIntegration : undefined}
-          readOnly={isReadOnly || isRunsMode || isFilesMode}
+          readOnly={isReadOnly || urlViewFlags.isRunsMode || urlViewFlags.isFilesMode}
           hasFitToViewRef={isRunsMode ? runsHasFitToViewRef : hasFitToViewRef}
           hasUserToggledSidebarRef={hasUserToggledSidebarRef}
           isSidebarOpenRef={isSidebarOpenRef}
